Diseño interfaz acorde a mockups
