+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
+        android:usesCleartextTraffic="true"
         android:theme="@style/Theme.Vinilos"
         tools:targetApi="31">
         <activity
