+++ app/src/main/java/com/misw/vinilos/ui/collectors/CollectorsFragment.kt
@@ -1,0 +1,28 @@
+package com.misw.vinilos.ui.collectors
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.misw.vinilos.databinding.FragmentCollectorsBinding
+
+/**
+ * Placeholder: sección Collectors se implementará más adelante.
+ */
+class CollectorsFragment : Fragment() {
+
+    private var _binding: FragmentCollectorsBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        _binding = FragmentCollectorsBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
+
