+++ app/src/main/res/color/vin_bottom_nav_item.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/vin_primary" android:state_checked="true" />
+    <item android:color="@color/vin_on_surface_muted" />
+</selector>
+
