+++ app/src/main/res/drawable/bg_header_overlay.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient
+        android:angle="90"
+        android:startColor="#CC0F1012"
+        android:centerColor="#660F1012"
+        android:endColor="#000F1012" />
+</shape>
