+++ app/src/main/res/layout/content_main.xml
@@ -3,6 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/vin_bg"
     app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
     <fragment
