+++ app/src/main/res/layout/fragment_performer_detail.xml
@@ -4,27 +4,40 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:background="@color/vin_bg"
     android:fillViewport="true">
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
         <ImageView
             android:id="@+id/performerImage"
-            android:layout_width="match_parent"
-            android:layout_height="250dp"
+            android:layout_width="0dp"
+            android:layout_height="@dimen/header_height"
             android:scaleType="centerCrop"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             tools:src="@tools:sample/backgrounds/scenic" />
+
+        <!-- Overlay para legibilidad del título sobre el header -->
+        <View
+            android:id="@+id/headerOverlay"
+            android:layout_width="0dp"
+            android:layout_height="@dimen/header_height"
+            android:background="@drawable/bg_header_overlay"
+            app:layout_constraintTop_toTopOf="@id/performerImage"
+            app:layout_constraintStart_toStartOf="@id/performerImage"
+            app:layout_constraintEnd_toEndOf="@id/performerImage" />
+
         <TextView
             android:id="@+id/performerName"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:textSize="24sp"
-            android:textStyle="bold"
-            app:layout_constraintTop_toBottomOf="@id/performerImage"
+            android:layout_marginStart="@dimen/space_screen"
+            android:layout_marginEnd="@dimen/space_screen"
+            android:layout_marginTop="@dimen/space_l"
+            style="@style/TextAppearance.Vinilos.ScreenTitle"
+            app:layout_constraintTop_toTopOf="@id/performerImage"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             tools:text="Rubén Blades" />
@@ -32,11 +45,11 @@
             android:id="@+id/performerDescription"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginTop="8dp"
-            android:textSize="16sp"
-            app:layout_constraintTop_toBottomOf="@id/performerName"
+            android:layout_marginStart="@dimen/space_screen"
+            android:layout_marginEnd="@dimen/space_screen"
+            android:layout_marginTop="@dimen/space_m"
+            style="@style/TextAppearance.Vinilos.Body"
+            app:layout_constraintTop_toBottomOf="@id/performerImage"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             tools:text="Descripción del actor" />
@@ -44,17 +57,20 @@
             android:id="@+id/tvAlbumsLabel"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_margin="16dp"
+            android:layout_marginStart="@dimen/space_screen"
+            android:layout_marginEnd="@dimen/space_screen"
+            android:layout_marginTop="@dimen/space_l"
             android:text="Álbumes"
-            android:textStyle="bold"
-            android:textSize="20sp"
+            style="@style/TextAppearance.Vinilos.Title"
             app:layout_constraintTop_toBottomOf="@id/performerDescription"
             app:layout_constraintStart_toStartOf="parent" />
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rvAlbums"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
+            android:layout_marginTop="@dimen/space_s"
+            android:layout_marginStart="@dimen/space_screen"
+            android:layout_marginEnd="@dimen/space_screen"
             android:nestedScrollingEnabled="false"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
             app:layout_constraintTop_toBottomOf="@id/tvAlbumsLabel"
