+++ app/src/main/res/layout/item_album.xml
@@ -1,30 +1,44 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:orientation="horizontal"
-    android:padding="8dp">
-
-    <ImageView
-        android:id="@+id/ivAlbumCover"
-        android:layout_width="64dp"
-        android:layout_height="64dp"
-        android:contentDescription="Carátula del álbum"/>
+    android:layout_margin="@dimen/space_s"
+    style="@style/Widget.Vinilos.Card">
 
     <LinearLayout
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:paddingStart="8dp">
+        android:orientation="vertical">
 
-        <TextView
-            android:id="@+id/tvAlbumName"
-            android:layout_width="wrap_content"
+        <ImageView
+            android:id="@+id/ivAlbumCover"
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/album_cover_grid_height_compact"
+            android:scaleType="centerCrop"
+            android:contentDescription="Carátula del álbum" />
+
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:textStyle="bold"/>
+            android:orientation="vertical"
+            android:padding="@dimen/space_s">
 
-        <TextView
-            android:id="@+id/tvAlbumGenre"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"/>
+            <TextView
+                android:id="@+id/tvAlbumName"
+                style="@style/TextAppearance.Vinilos.Title"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:maxLines="1"
+                android:ellipsize="end" />
+
+            <TextView
+                android:id="@+id/tvAlbumGenre"
+                style="@style/TextAppearance.Vinilos.Accent"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="@dimen/space_xs"
+                android:maxLines="1"
+                android:ellipsize="end" />
+        </LinearLayout>
     </LinearLayout>
-</LinearLayout>
+</com.google.android.material.card.MaterialCardView>
