+++ app/src/main/res/layout/item_track.xml
@@ -2,20 +2,26 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:padding="16dp"
-    android:orientation="horizontal">
+    android:orientation="horizontal"
+    android:paddingStart="@dimen/space_m"
+    android:paddingEnd="@dimen/space_m"
+    android:paddingTop="@dimen/space_s"
+    android:paddingBottom="@dimen/space_s">
+
     <TextView
         android:id="@+id/trackName"
+        style="@style/TextAppearance.Vinilos.Body"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_weight="1"
-        android:textSize="16sp"
-        android:textColor="@android:color/black"/>
+        android:maxLines="1"
+        android:ellipsize="end" />
+
     <TextView
         android:id="@+id/trackDuration"
+        style="@style/TextAppearance.Vinilos.Meta"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textSize="14sp"
-        android:textColor="@android:color/darker_gray"
-        android:layout_marginStart="8dp"/>
+        android:layout_marginStart="@dimen/space_s" />
+
 </LinearLayout>
