+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/AlbumListFragment"
+        android:title="Albums"
+        android:icon="@android:drawable/ic_media_play" />
+
+    <item
+        android:id="@+id/PerformerListFragment"
+        android:title="Artists"
+        android:icon="@android:drawable/ic_menu_myplaces" />
+
+    <item
+        android:id="@+id/CollectorsFragment"
+        android:title="Collectors"
+        android:icon="@android:drawable/ic_menu_agenda" />
+
+</menu>
+
