+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,19 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
-    <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.misw.vinilos.FirstFragment"
-        android:label="First Fragment"
-        tools:layout="@layout/fragment_first">
-        <action
-            android:id="@+id/action_FirstFragment_to_AlbumListFragment"
-            app:destination="@id/AlbumListFragment" />
-        <action
-            android:id="@+id/action_FirstFragment_to_PerformerListFragment"
-            app:destination="@id/PerformerListFragment" />
-    </fragment>
+    app:startDestination="@id/AlbumListFragment">
     <fragment
         android:id="@+id/AlbumListFragment"
         android:name="com.misw.vinilos.ui.albums.AlbumListFragment"
@@ -61,4 +49,11 @@
             android:id="@+id/action_PerformerDetailFragment_to_AlbumDetailFragment"
             app:destination="@id/AlbumDetailFragment" />
     </fragment>
+
+    <!-- Placeholder para Collectors (se implementará más adelante) -->
+    <fragment
+        android:id="@+id/CollectorsFragment"
+        android:name="com.misw.vinilos.ui.collectors.CollectorsFragment"
+        android:label="Collectors"
+        tools:layout="@layout/fragment_collectors" />
 </navigation>
