+++ app/src/main/res/values-night/themes.xml
@@ -1,7 +1,16 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.Vinilos" parent="Theme.Material3.DayNight.NoActionBar">
-        <!-- Customize your dark theme here. -->
-        <!-- <item name="colorPrimary">@color/my_dark_primary</item> -->
+    <style name="Base.Theme.Vinilos" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+        <!-- Mockup: Dark theme tokens (mantener en night para consistencia) -->
+        <item name="colorPrimary">@color/vin_primary</item>
+        <item name="colorSecondary">@color/vin_secondary</item>
+        <item name="android:colorBackground">@color/vin_bg</item>
+        <item name="colorSurface">@color/vin_surface</item>
+        <item name="colorOnSurface">@color/vin_on_surface</item>
+        <item name="colorOutline">@color/vin_outline</item>
+
+        <item name="android:statusBarColor">@color/vin_bg</item>
+        <item name="android:navigationBarColor">@color/vin_bg</item>
+        <item name="android:windowLightStatusBar">false</item>
     </style>
 </resources>
