+++ app/src/main/res/values/styles.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <!-- Typography -->
+    <style name="TextAppearance.Vinilos.ScreenTitle" parent="TextAppearance.MaterialComponents.Headline6">
+        <item name="android:textColor">@color/vin_on_surface</item>
+        <item name="android:textStyle">bold</item>
+    </style>
+
+    <style name="TextAppearance.Vinilos.Title" parent="TextAppearance.MaterialComponents.Subtitle1">
+        <item name="android:textColor">@color/vin_on_surface</item>
+        <item name="android:textStyle">bold</item>
+    </style>
+
+    <style name="TextAppearance.Vinilos.Body" parent="TextAppearance.MaterialComponents.Body1">
+        <item name="android:textColor">@color/vin_on_surface</item>
+    </style>
+
+    <style name="TextAppearance.Vinilos.Meta" parent="TextAppearance.MaterialComponents.Body2">
+        <item name="android:textColor">@color/vin_on_surface_muted</item>
+    </style>
+
+    <style name="TextAppearance.Vinilos.Accent" parent="TextAppearance.MaterialComponents.Body2">
+        <item name="android:textColor">@color/vin_secondary</item>
+        <item name="android:textStyle">bold</item>
+    </style>
+
+    <!-- Cards -->
+    <style name="Widget.Vinilos.Card" parent="Widget.MaterialComponents.CardView">
+        <item name="cardCornerRadius">@dimen/radius_l</item>
+        <item name="cardUseCompatPadding">true</item>
+        <item name="cardBackgroundColor">@color/vin_surface</item>
+    </style>
+
+    <!-- Shapes -->
+    <style name="ShapeAppearance.Vinilos" parent="ShapeAppearance.MaterialComponents.MediumComponent" />
+
+    <style name="ShapeAppearance.Vinilos.Circle">
+        <item name="cornerFamily">rounded</item>
+        <item name="cornerSize">50%</item>
+    </style>
+
+</resources>
+
+
