+++ app/src/main/res/values/themes.xml
@@ -1,9 +1,24 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.Vinilos" parent="Theme.Material3.DayNight.NoActionBar">
-        <!-- Customize your light theme here. -->
-        <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+    <style name="Base.Theme.Vinilos" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+        <!-- Mockup: Dark theme tokens -->
+        <item name="colorPrimary">@color/vin_primary</item>
+        <item name="colorSecondary">@color/vin_secondary</item>
+        <item name="android:colorBackground">@color/vin_bg</item>
+        <item name="colorSurface">@color/vin_surface</item>
+        <item name="colorOnSurface">@color/vin_on_surface</item>
+        <item name="colorOutline">@color/vin_outline</item>
+
+        <item name="android:statusBarColor">@color/vin_bg</item>
+        <item name="android:navigationBarColor">@color/vin_bg</item>
+        <item name="android:windowLightStatusBar">false</item>
     </style>
 
     <style name="Theme.Vinilos" parent="Base.Theme.Vinilos" />
+
+    <style name="Widget.Vinilos.Toolbar" parent="Widget.MaterialComponents.Toolbar">
+        <item name="android:background">@color/vin_bg</item>
+        <item name="titleTextColor">@color/vin_on_surface</item>
+        <item name="subtitleTextColor">@color/vin_on_surface_muted</item>
+    </style>
 </resources>
