+++ gradle.properties
@@ -21,4 +21,4 @@ kotlin.code.style=official
 # resources declared in the library itself and none from the library's dependencies,
 # thereby reducing the size of the R class for that library
 android.nonTransitiveRClass=true
-BASE_URL=https://back-vynils-heroku.herokuapp.com/
+BASE_URL=http://192.168.2.14:3000/
